move encryption acceptance tests to core tests
