+++ src/NServiceBus.AcceptanceTests/Core/Encryption/When_using_Rijndael_with_config.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Encryption
+﻿namespace NServiceBus.AcceptanceTests.Core.Encryption
 {
     using System;
     using System.Collections.Generic;
@@ -89,7 +89,7 @@
             }
         }
 
-        
+
         public class MessageWithSecretData : IMessage
         {
             public WireEncryptedString Secret { get; set; }
@@ -97,14 +97,14 @@
             public List<CreditCardDetails> CreditCards { get; set; }
         }
 
-        
+
         public class CreditCardDetails
         {
             public DateTime ValidTo { get; set; }
             public WireEncryptedString Number { get; set; }
         }
 
-        
+
         public class MySecretSubProperty
         {
             public WireEncryptedString Secret { get; set; }
