+++ src/NServiceBus.AcceptanceTests/Core/Encryption/When_using_Rijndael_with_custom.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Encryption
+﻿namespace NServiceBus.AcceptanceTests.Core.Encryption
 {
     using System;
     using System.Collections.Generic;
