+++ src/NServiceBus.AcceptanceTests/Core/Encryption/When_using_Rijndael_with_multikey.cs
@@ -1,21 +1,20 @@
-﻿namespace NServiceBus.AcceptanceTests.Encryption
+﻿namespace NServiceBus.AcceptanceTests.Core.Encryption
 {
     using System.Collections.Generic;
     using System.Text;
     using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
-    using MessageMutator;
     using NUnit.Framework;
     using ScenarioDescriptors;
 
-    public class When_using_Rijndael_without_incoming_key_identifier : NServiceBusAcceptanceTest
+    public class When_using_Rijndael_with_multikey : NServiceBusAcceptanceTest
     {
         [Test]
-        public Task Should_process_decrypted_message_without_key_identifier()
+        public Task Should_receive_decrypted_message()
         {
             return Scenario.Define<Context>()
-                .WithEndpoint<Sender>(b => b.When((bus, context) => bus.Send(new MessageWithSecretData
+                .WithEndpoint<Sender>(b => b.When((session, context) => session.Send(new MessageWithSecretData
                 {
                     Secret = "betcha can't guess my secret"
                 })))
@@ -36,7 +35,7 @@
         {
             public Sender()
             {
-                EndpointSetup<DefaultServer>(builder => builder.RijndaelEncryptionService("will-be-removed-by-transport-mutator", Encoding.ASCII.GetBytes("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")))
+                EndpointSetup<DefaultServer>(builder => builder.RijndaelEncryptionService("1st", Encoding.ASCII.GetBytes("gdDbqRpqdRbTs3mhdZh9qCaDaxJXl+e6")))
                     .AddMapping<MessageWithSecretData>(typeof(Receiver));
             }
         }
@@ -45,16 +44,18 @@
         {
             public Receiver()
             {
+                var key = Encoding.ASCII.GetBytes("gdDbqRpqdRbTs3mhdZh9qCaDaxJXl+e6");
                 var keys = new Dictionary<string, byte[]>
                 {
-                    {"new", Encoding.ASCII.GetBytes("bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb")}
+                    {"2nd", Encoding.ASCII.GetBytes("gdDbqRpqdRbTs3mhdZh9qCaDaxJXl+e6")},
+                    {"1st", key}
                 };
 
                 var expiredKeys = new[]
                 {
-                    Encoding.ASCII.GetBytes("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")
+                    key
                 };
-                EndpointSetup<DefaultServer>(builder => builder.RijndaelEncryptionService("new", keys, expiredKeys));
+                EndpointSetup<DefaultServer>(builder => builder.RijndaelEncryptionService("2nd", keys, expiredKeys));
             }
 
             public class Handler : IHandleMessages<MessageWithSecretData>
@@ -65,6 +66,7 @@
                 {
                     Context.Secret = message.Secret.Value;
                     Context.Done = true;
+
                     return Task.FromResult(0);
                 }
             }
@@ -74,20 +76,6 @@
         public class MessageWithSecretData : IMessage
         {
             public WireEncryptedString Secret { get; set; }
-        }
-
-        class RemoveKeyIdentifierHeaderMutator : IMutateIncomingTransportMessages, INeedInitialization
-        {
-            public Task MutateIncoming(MutateIncomingTransportMessageContext context)
-            {
-                context.Headers.Remove(Headers.RijndaelKeyIdentifier);
-                return Task.FromResult(0);
-            }
-
-            public void Customize(EndpointConfiguration configuration)
-            {
-                configuration.RegisterComponents(c => c.ConfigureComponent<RemoveKeyIdentifierHeaderMutator>(DependencyLifecycle.InstancePerCall));
-            }
         }
     }
 }
