+++ src/NServiceBus.AcceptanceTests/Core/Encryption/When_using_Rijndael_with_unobtrusive_mode.cs
@@ -1,4 +1,4 @@
-namespace NServiceBus.AcceptanceTests.Encryption
+namespace NServiceBus.AcceptanceTests.Core.Encryption
 {
     using System;
     using System.Collections.Generic;
