+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -70,7 +70,7 @@
     <Compile Include="DataBus\When_sending_databus_properties_with_unobtrusive.cs" />
     <Compile Include="Basic\When_sending_interface_message_with_conventions.cs" />
     <Compile Include="Core\DelayedDelivery\TimeoutManager\When_using_external_timeout_manager.cs" />
-    <Compile Include="Encryption\When_using_Rijndael_with_unobtrusive_mode.cs" />
+    <Compile Include="Core\Encryption\When_using_Rijndael_with_unobtrusive_mode.cs" />
     <Compile Include="Forwarding\When_requesting_message_to_be_forwarded.cs" />
     <Compile Include="Performance\TimeToBeReceived\When_TimeToBeReceived_used_with_unobtrusive_mode.cs" />
     <Compile Include="Recoverability\Retries\When_custom_policy_provided.cs" />
@@ -199,7 +199,7 @@
     <Compile Include="Reliability\Outbox\When_a_message_is_audited.cs" />
     <Compile Include="Reliability\Outbox\When_clearing_saga_timeouts.cs" />
     <Compile Include="Core\Outbox\When_dispatching_transport_operations.cs" />
-    <Compile Include="Encryption\When_using_Rijndael_without_incoming_key_identifier.cs" />
+    <Compile Include="Core\Encryption\When_using_Rijndael_without_incoming_key_identifier.cs" />
     <Compile Include="PerfMon\CriticalTime\When_deferring_a_message.cs" />
     <Compile Include="Audit\When_audit_is_overridden_in_code.cs" />
     <Compile Include="Reliability\Outbox\When_a_duplicate_message_arrives.cs" />
@@ -266,9 +266,9 @@
     <Compile Include="Core\UnitOfWork\TransactionScope\When_using_timeout_greater_than_machine_max.cs" />
     <Compile Include="Core\UnitOfWork\TransactionScope\When_used_with_transport_scopes.cs" />
     <Compile Include="Core\UnitOfWork\When_a_subscription_message_arrives.cs" />
-    <Compile Include="Encryption\When_using_Rijndael_with_multikey.cs" />
-    <Compile Include="Encryption\When_using_Rijndael_with_custom.cs" />
-    <Compile Include="Encryption\When_using_Rijndael_with_config.cs" />
+    <Compile Include="Core\Encryption\When_using_Rijndael_with_multikey.cs" />
+    <Compile Include="Core\Encryption\When_using_Rijndael_with_custom.cs" />
+    <Compile Include="Core\Encryption\When_using_Rijndael_with_config.cs" />
     <Compile Include="Mutators\Issue_1980.cs" />
     <Compile Include="Mutators\When_outgoing_mutator_replaces_instance.cs" />
     <Compile Include="Mutators\When_defining_outgoing_message_mutators.cs" />
@@ -291,7 +291,7 @@
     <Compile Include="Basic\When_multiple_mappings_exists.cs" />
     <Compile Include="Correlation\When_using_a_custom_correlation_id.cs" />
     <Compile Include="Config\When_only_abstract_config_override_is_found.cs" />
-    <Compile Include="Encryption\When_using_encryption_with_custom_service.cs" />
+    <Compile Include="Core\Encryption\When_using_encryption_with_custom_service.cs" />
     <Compile Include="Reliability\Outbox\When_sending_from_a_non_dtc_endpoint.cs" />
     <Compile Include="Reliability\Outbox\When_receiving_a_message.cs" />
     <Compile Include="NServiceBusAcceptanceTest.cs" />
